JDBC Connector fix column sizes of types
Implement COLUMN_SIZE, DECIMAL_DIGITS and CHAR_OCTET_LENGTH for
following types.

* BIGINT, INTEGER, SMALLINT, TINYINT, DOUBLE, VARCHAR, VARBINARY, TIME,
  TIME WITH TIME ZONE, TIMESTAMP, TIMESTAMP WITH TIME ZONE, DATE,
  DECIMAL.

Before this commit getColumns() always returns 0 for COLUMN_SIZE and
CHAR_OCTET_LENGTH, and returns 19 for DECIMAL_DIGITS when a type is
BIGINT. For BIGINT, DECIMAL_DIGITS == scale should be 0, not 19.

So this commit is worth applying I think, but so far I've not yet
tested any JDBC Drivers.