Floating point's COLUMN_SIZE and NUM_PREC_RADIX
Return proper values for COLUMN_SIZE and NUM_PREC_RADIX in case where a
type is FLOAT or DOUBLE. In Presto FLOAT and DOUBLE are IEEE 754
single/double-precision floating number and precisions are 53 and 24
bits.

This commit also changes the value of NUM_PREC_RADIX in case where a
type is not for number. It no longer returns 10 for non number type like
String and Date but returns NULL.