DOUBLE's COLUMN_SIZE and NUM_PREC_RADIX
Return proper values for COLUMN_SIZE and NUM_PREC_RADIX in case where a
type is DOUBLE. In Presto DOUBLE is IEEE 754 double-precision floating
number and precision is 53 bit.

This commit also changes the value of NUM_PREC_RADIX in case where a
type is not for number. It no longer returns 10 for non number type like
String and Date but returns NULL.