Optimize geometry_to_bing_tiles function
Optimized the algorithm for identifying a set of covering tiles for zoom levels > 10. The idea is to identify large tiles which are fully covered by the geometry first. For each such tile, we can cheaply compute all the containing tiles at the right zoom level and append them to results in bulk. This way we perform a single containment check instead of 2 to the power of level delta intersection checks, where level delta is the difference between the desired zoom level and level of the large tile covered by the geometry.